+++ Commands/sparkle_help/reporting_scenario.py
@@ -158,8 +158,6 @@ class ReportingScenario:
 			self.__init_section(section)
 			self.__scenario[section][name] = str(value)
 
-		self.write_scenario_ini()
-
 		return
 
 
@@ -183,8 +181,6 @@ class ReportingScenario:
 		if value != None:
 			self.__init_section(section)
 			self.__scenario[section][name] = value.name
-
-		self.write_scenario_ini()
 
 		return