+++ Commands/test/validate_configured_vs_default.sh
@@ -24,18 +24,23 @@ cp $slurm_settings_test $slurm_settings_path # Activate test settings
 sparkle_test_settings_path="Commands/test/test_files/sparkle_settings.ini"
 
 # Prepare for test
-instances_path_train="Examples/Resources/Instances/PTN/"
-instances_path_test="Examples/Resources/Instances/PTN2/"
-solver_path="Examples/Resources/Solvers/PbO-CCSAT-Generic/"
+examples_path="Examples/Resources/"
+instances_path_train="Instances/PTN/"
+instances_path_test="Instances/PTN2/"
+solver_path="Solvers/PbO-CCSAT-Generic/"
+instances_src_path_train="${examples_path}${instances_path_train}"
+instances_src_path_test="${examples_path}${instances_path_test}"
+solver_src_path="${examples_path}${solver_path}"
+
 configuration_results_path="Commands/test/test_files/results/"
 configuration_files_path="Commands/test/test_files/PbO-CCSAT-Generic/PTN_train.txt"
 smac_path="Components/smac-v2.10.03-master-778/"
 smac_configuration_files_path="$smac_path/example_scenarios/PbO-CCSAT-Generic/"
 
 Commands/initialise.py > /dev/null
-Commands/add_instances.py $instances_path_train > /dev/null
-Commands/add_instances.py $instances_path_test > /dev/null
-Commands/add_solver.py --deterministic 0 $solver_path > /dev/null
+Commands/add_instances.py $instances_src_path_train > /dev/null
+Commands/add_instances.py $instances_src_path_test > /dev/null
+Commands/add_solver.py --deterministic 0 $solver_src_path > /dev/null
 
 # Copy configuration results and other files to simulate the configuration command
 cp -r $configuration_results_path $smac_path
@@ -46,7 +51,7 @@ cp $configuration_files_path $smac_configuration_files_path
 output=$(Commands/validate_configured_vs_default.py --solver $solver_path --instance-set-train $instances_path_train --instance-set-test $instances_path_test --settings-file $sparkle_test_settings_path | tail -1)
 output_true="c Running validation in parallel. Waiting for Slurm job with id: "
 
-if [[ $output =~ [^$output_true] ]];
+if [[ $output =~ "${output_true}" ]];
 then
 	echo "[success] validate_configured_vs_default with both train and test sets test succeeded"
     jobid=${output##* }
@@ -60,7 +65,7 @@ fi
 # Test configured solver and default solver with just training set
 output=$(Commands/validate_configured_vs_default.py --solver $solver_path --instance-set-train $instances_path_train --settings-file $sparkle_test_settings_path | tail -1)
 
-if [[ $output =~ [^$output_true] ]];
+if [[ $output =~ "${output_true}" ]];
 then
 	echo "[success] validate_configured_vs_default with just training set test succeeded"
     jobid=${output##* }